backend/requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.14.0
djangorestframework-simplejwt>=5.2.2
PyJWT>=2.8.0
python_dotenv>=1.1.0
django-cors-headers>=4.7.0
google-auth>=2.40.3